+++ .vscode/launch.json
@@ -6,7 +6,7 @@
             "type": "go",
             "request": "launch",
             "mode": "auto",
-            "program": "${workspaceFolder}/pkg/main.go",
+            "program": "${workspaceFolder}/pkg/",
             "env": {},
             "args": [
                 "--standalone=true",
