+++ go.mod
@@ -3,15 +3,13 @@ module github.com/grafana/iot-sitewise-datasource
 go 1.16
 
 require (
-	github.com/aws/aws-sdk-go v1.38.22
-	github.com/grafana/grafana-aws-sdk v0.5.0
-	github.com/grafana/grafana-plugin-sdk-go v0.92.0
+	github.com/aws/aws-sdk-go v1.38.39
+	github.com/grafana/grafana-aws-sdk v0.6.0
+	github.com/grafana/grafana-plugin-sdk-go v0.97.0
 	github.com/magefile/mage v1.11.0
 	github.com/patrickmn/go-cache v2.1.0+incompatible
 	github.com/pkg/errors v0.9.1
-	github.com/prometheus/client_golang v1.10.0 // indirect
 	github.com/stretchr/objx v0.3.0 // indirect
 	github.com/stretchr/testify v1.7.0
-	github.com/vektra/mockery/v2 v2.7.4 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
