+++ pkg/main.go
@@ -3,17 +3,14 @@ package main
 import (
 	"os"
 
-	"github.com/grafana/grafana-plugin-sdk-go/backend"
-	"github.com/grafana/grafana-plugin-sdk-go/experimental"
-	"github.com/grafana/iot-sitewise-datasource/pkg/plugin"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/datasource"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
+	"github.com/grafana/iot-sitewise-datasource/pkg/server"
 )
 
 func main() {
-	err := experimental.DoGRPC("sitewise-datasource", plugin.GetDatasourceServeOpts())
-
-	// Log any error if we could start the plugin.
-	if err != nil {
-		backend.Logger.Error(err.Error())
+	if err := datasource.Manage("sitewise-datasource", server.NewServerInstance, datasource.ManageOpts{}); err != nil {
+		log.DefaultLogger.Error(err.Error())
 		os.Exit(1)
 	}
 }
