+++ pkg/resource/caching_resource_provider_test.go
@@ -2,12 +2,13 @@ package resource
 
 import (
 	"context"
+	"testing"
+
 	"github.com/grafana/iot-sitewise-datasource/pkg/sitewise/client/mocks"
 	"github.com/grafana/iot-sitewise-datasource/pkg/testdata"
 	"github.com/patrickmn/go-cache"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
-	"testing"
 )
 
 func tdpath(filename string) string {
