+++ pkg/server/server.go
@@ -15,25 +15,24 @@ import (
 )
 
 type Server struct {
-	Datasource Datasource
+	Datasource    Datasource
+	channelPrefix string
+	closeCh       chan struct{}
 }
+
+// Make sure SampleDatasource implements required interfaces.
+// This is important to do since otherwise we will only get a
+// not implemented error response from plugin in runtime.
+var (
+	_ backend.QueryDataHandler      = (*Server)(nil)
+	_ backend.CheckHealthHandler    = (*Server)(nil)
+	_ instancemgmt.InstanceDisposer = (*Server)(nil)
+)
 
 // QueryHandlerFunc is the function signature used for mux.HandleFunc
 // Looks like mux.HandleFunc uses backend.QueryHandlerFunc
 // type QueryDataHandlerFunc func(ctx context.Context, req *QueryDataRequest) (*QueryDataResponse, error)
 type QueryHandlerFunc func(context.Context, *backend.QueryDataRequest, backend.DataQuery) backend.DataResponse
-
-func processQueries(ctx context.Context, req *backend.QueryDataRequest, handler QueryHandlerFunc) *backend.QueryDataResponse {
-	res := backend.Responses{}
-	for _, v := range req.Queries {
-		res[v.RefID] = handler(ctx, req, v)
-	}
-
-	return &backend.QueryDataResponse{
-		Responses: res,
-	}
-
-}
 
 func DataResponseErrorUnmarshal(err error) backend.DataResponse {
 	return backend.DataResponse{
@@ -71,4 +70,34 @@ func NewServerInstance(settings backend.DataSourceInstanceSettings) (instancemgm
 		Datasource: ds,
 	}
 	return srvr, nil
+}
+
+// QueryData handles multiple queries and returns multiple responses.
+// req contains the queries []DataQuery (where each query contains RefID as a unique identifer).
+// The QueryDataResponse contains a map of RefID to the response for each query, and each response
+// contains Frames ([]*Frame).
+func (s *Server) QueryData(ctx context.Context, req *backend.QueryDataRequest) (*backend.QueryDataResponse, error) {
+	m := GetQueryHandlers(s)
+	return m.QueryData(ctx, req)
+}
+
+// CheckHealth handles health checks sent from Grafana to the plugin.
+// The main use case for these health checks is the test button on the
+// datasource configuration page which allows users to verify that
+// a datasource is working as expected.
+func (s *Server) CheckHealth(ctx context.Context, req *backend.CheckHealthRequest) (*backend.CheckHealthResult, error) {
+	if err := s.Datasource.HealthCheck(ctx, req); err != nil {
+		return &backend.CheckHealthResult{
+			Status:  backend.HealthStatusError,
+			Message: err.Error(),
+		}, nil
+	}
+	return &backend.CheckHealthResult{
+		Status:  backend.HealthStatusOk,
+		Message: backend.HealthStatusOk.String(),
+	}, nil
+}
+
+func (s *Server) Dispose() {
+	close(s.closeCh)
 }
